% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_test.R
\name{TMSapp.check}
\alias{TMSapp.check}
\title{TMSBrianApp data check}
\usage{
TMSapp.check(TMSdata)
}
\arguments{
\item{TMSdata}{A data frame to be checked for eligibility.}
}
\value{
Returns a data frame if minimum requirement test is passed.
}
\description{
Check data for minimum requirement for TMSBrainApp analysis.
}
\examples{
\dontrun{
TMSapp.check(TMSdata)
}

}
