% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{TMSapp.run}
\alias{TMSapp.run}
\title{A shiny Web interface for hotspot finding in a TMS brain mapping data}
\usage{
TMSapp.run(...)
}
\value{
Returns the location of the hotspot and the posterior summary of this location on a webpage. Additionally,
several hypothesis can be tested on the hotspot using Monte Carlo samples from the posterior distribution.
}
\description{
Accept TMS data and make statistical inferences about the hotspot
 location using Bayesian spatial statistical analysis.  From the running web page,
  fixed and random effects can be passed into the model.
  It is recommended that user data is checked and extracted using
   the function TMSapp.check().
}
\examples{
\dontrun{
TMSapp.run()
}

}
\seealso{
\code{\link{SpatialEpiApp}}
}
